A 32bit integer is big enough to cover all byte address of m210.
  * Packets are numbered with 16 bit integers.
    => Maximum number of packets: 2**16 = 65536

  * Each packet is 64 bytes wide, last 62 bytes represent bytes in memory.
    The first two bytes represent the packet sequence number.
    => Maximum number of bytes in memory: 2**16 * 62 = 4063232

  * A 32bit integer can address 2**32 different bytes which is way more
    than maximum number of bytes in m210 memory.

+++ src/m210.c
@@ -476,7 +476,7 @@ enum m210_err m210_read_packet(const struct m210 *m210, struct m210_packet *pack
     return err_ok;
 }
 
-enum m210_err m210_get_notes_size(const struct m210 *m210, ssize_t *size)
+enum m210_err m210_get_notes_size(const struct m210 *m210, uint32_t *size)
 {
     enum m210_err err;
     uint16_t packet_count;
